+++ R/get_pvol.R
@@ -135,6 +135,7 @@ select_get_pvol_function <- function(radar, ..., call = rlang::caller_env()) {
     cntry_code == "cz" ~ "get_pvol_cz",
     cntry_code == "se" ~ "get_pvol_se",
     cntry_code == "ro" ~ "get_pvol_ro",
+    cntry_code == "sk" ~ "get_pvol_sk",
     .default = NA
   ))
   if (rlang::is_na(fun)) {