+++ tests/testthat/teardown_remove_cache.R
@@ -1,0 +1,4 @@
+unlink(
+  tools::R_user_dir("getRad", "cache"),
+  recursive = TRUE
+)